Improve logging of integration tests
