+++ test/init_test.go
@@ -1,5 +1,3 @@
-// +build e2e
-
 package test
 
 import (
