+++ test/pipelineRun_conditions.go
@@ -7,8 +7,10 @@ import (
 	"github.com/SAP/stewardci-core/pkg/k8s"
 )
 
+// PipelineRunCheck is a check for a PipelineRun
 type PipelineRunCheck func(k8s.PipelineRun) bool
 
+// CreatePipelineRunCondition returns a WaitCondition for a pipelineRun with a dedicated PipelineCheck
 func CreatePipelineRunCondition(pipelineRun *api.PipelineRun, check PipelineRunCheck, desc string) WaitCondition {
 	return NewWaitCondition(func(clientFactory k8s.ClientFactory) (bool, error) {
 		fetcher := k8s.NewPipelineRunFetcher(clientFactory)
@@ -22,8 +24,10 @@ func CreatePipelineRunCondition(pipelineRun *api.PipelineRun, check PipelineRunC
 		fmt.Sprintf("PRC_%s_%s_%s", pipelineRun.GetNamespace(), pipelineRun.GetName(), desc))
 }
 
+// PipelineRunHasStateResult returns a PipelineRunCheck which checks if a pipelineRun has a dedicated result
 func PipelineRunHasStateResult(result api.Result) PipelineRunCheck {
 	return func(pr k8s.PipelineRun) bool {
+
 		return pr.GetStatus().Result == result
 	}
 }
