+++ test/tenant_conditions.go
@@ -7,8 +7,10 @@ import (
 	"github.com/SAP/stewardci-core/pkg/k8s"
 )
 
+// TenantCheck is a check for a tenant
 type TenantCheck func(*api.Tenant) bool
 
+// CreateTenantCondition creates a WaitCondition for a tenant with a dedicated check
 func CreateTenantCondition(tenant *api.Tenant, check TenantCheck, desc string) WaitCondition {
 	key := fmt.Sprintf("%s/%s", tenant.GetNamespace(), tenant.GetName())
 	return NewWaitCondition(func(clientFactory k8s.ClientFactory) (bool, error) {
@@ -22,6 +24,7 @@ func CreateTenantCondition(tenant *api.Tenant, check TenantCheck, desc string) W
 		fmt.Sprintf("TenantCondition_%s_%s", key, desc))
 }
 
+// TenantHasStateResult creates a TenantCheck which checks for a dedicated State
 func TenantHasStateResult(result api.TenantResult) TenantCheck {
 	return func(tenant *api.Tenant) bool {
 		return tenant.Status.Result == result
