+++ test/wait.go
@@ -15,6 +15,7 @@ const (
 	timeout  = 2 * time.Minute
 )
 
+// Waiter is a waiter waiting for a condition to be fullfilled
 type Waiter interface {
 	WaitFor(condition WaitCondition) error
 }
@@ -23,22 +24,30 @@ type waiter struct {
 	clientFactory k8s.ClientFactory
 }
 
+// NewWaiter returns a new Waiter
 func NewWaiter(clientFactory k8s.ClientFactory) Waiter {
 	return &waiter{clientFactory: clientFactory}
 }
 
+// WaitFor waits for a condition
+// it returns an error if condition is not fullfilled
 func (w *waiter) WaitFor(condition WaitCondition) error {
+	log.Printf("wait for %s", condition.Name())
+	startTime := time.Now()
 	_, span := trace.StartSpan(context.Background(), condition.Name())
 	defer span.End()
-	return wait.PollImmediate(interval, timeout, func() (bool, error) {
-		return condition.Wait(w.clientFactory)
+
+	err := wait.PollImmediate(interval, timeout, func() (bool, error) {
+		return condition.Check(w.clientFactory)
 	})
+	log.Printf("waiting completed for %s after %s", condition.Name(), time.Now().Sub(startTime))
+	return err
 }
 
-func (w *waiter) MyWaitFor(condition WaitCondition) error {
+func (w *waiter) myWaitFor(condition WaitCondition) error {
 	time.Sleep(interval)
 	for {
-		result, err := condition.Wait(w.clientFactory)
+		result, err := condition.Check(w.clientFactory)
 		log.Printf("MyWaitFor: %t, %s", result, err)
 		if err != nil {
 			return err
